@@ -1,5 +1,5 @@
 {
-  "name": "mtdd",
+  "name": "vue-test",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,30 +8,17 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@tailwindcss/postcss7-compat": "^2.1.0",
-    "autoprefixer": "^9.8.6",
     "core-js": "^3.6.5",
-    "headlessui": "^0.0.0",
-    "heroicons": "^1.0.1",
-    "postcss": "^7.0.35",
-    "register-service-worker": "^1.7.1",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.1.0",
-    "vue": "^2.6.11"
+    "vue": "^3.0.0"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-pwa": "~4.5.0",
-    "@vue/cli-service": "^4.1.1",
-    "@vue/eslint-config-standard": "^5.1.2",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "eslint-plugin-vue": "^7.0.0"
   },
   "eslintConfig": {
     "root": true,
@@ -39,9 +26,8 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended",
-      "@vue/standard"
+      "plugin:vue/vue3-essential",
+      "eslint:recommended"
     ],
     "parserOptions": {
       "parser": "babel-eslint"
